[AIRFLOW-5201] Move GCP Functions to core
This commit moves GCP Functions from contrib to core.
For more information check AIP-21.
