+++ tests/gcp/operators/test_functions.py
@@ -24,7 +24,7 @@ from copy import deepcopy
 from googleapiclient.errors import HttpError
 from parameterized import parameterized
 
-from airflow.contrib.operators.gcp_function_operator import \
+from airflow.gcp.operators.functions import \
     GcfFunctionDeployOperator, GcfFunctionDeleteOperator, FUNCTION_NAME_PATTERN
 from airflow import AirflowException
 from airflow.version import version
@@ -71,7 +71,7 @@ def _prepare_test_bodies():
 
 class GcfFunctionDeployTest(unittest.TestCase):
     @parameterized.expand(_prepare_test_bodies())
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_body_empty_or_missing_fields(self, body, message, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         with self.assertRaises(AirflowException) as cm:
@@ -88,7 +88,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
                                           gcp_conn_id='google_cloud_default')
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_deploy_execute(self, mock_hook):
         mock_hook.return_value.get_function.side_effect = mock.Mock(
             side_effect=HttpError(resp=MOCK_RESP_404, content=b'not found'))
@@ -115,7 +115,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
             body=expected_body
         )
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_update_function_if_exists(self, mock_hook):
         mock_hook.return_value.get_function.return_value = True
         mock_hook.return_value.update_function.return_value = True
@@ -140,7 +140,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
             expected_body, expected_body.keys())
         mock_hook.return_value.create_new_function.assert_not_called()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_empty_project_id_is_ok(self, mock_hook):
         operator = GcfFunctionDeployOperator(
             location="test_region",
@@ -160,7 +160,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
             location="test_region",
             body=new_body)
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_empty_location(self, mock_hook):
         with self.assertRaises(AirflowException) as cm:
             GcfFunctionDeployOperator(
@@ -174,7 +174,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
         mock_hook.assert_called_once_with(api_version='v1',
                                           gcp_conn_id='google_cloud_default')
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_empty_body(self, mock_hook):
         with self.assertRaises(AirflowException) as cm:
             GcfFunctionDeployOperator(
@@ -191,7 +191,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
     @parameterized.expand([
         (runtime,) for runtime in VALID_RUNTIMES
     ])
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_correct_runtime_field(self, runtime, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -215,7 +215,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
             "projects/PRÓJECT/global/networks/netwórk-01"
         ]
     ])
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_valid_network_field(self, network, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -238,7 +238,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
             {"label_324234_a_b_c": 'value-01_93'},
         ]
     ])
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_valid_labels_field(self, labels, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -254,7 +254,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
                                           gcp_conn_id='google_cloud_default')
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_validation_disabled(self, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = {
@@ -273,7 +273,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
                                           gcp_conn_id='google_cloud_default')
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_body_validation_simple(self, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -309,7 +309,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
         ('maxInstances', '-1', "The max instances parameter has to be greater than 0"),
         ('maxInstances', 'ss', "invalid literal for int() with base 10: 'ss'"),
     ])
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_invalid_field_values(self, key, value, message, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -362,7 +362,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
          "The body field 'source_code.sourceRepository.url' of value '' does not match"),
     ]
     )
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_invalid_source_code_union_field(self, source_code, message, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         body = deepcopy(VALID_BODY)
@@ -397,7 +397,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
           'repos/b/revisions/c/paths/d'}},
          'test_project_id'),
     ])
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_valid_source_code_union_field(self, source_code, project_id, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         mock_hook.return_value.get_function.side_effect = mock.Mock(
@@ -463,7 +463,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
          "should be of dictionary type")
     ]
     )
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_invalid_trigger_union_field(self, trigger, message, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         body = deepcopy(VALID_BODY)
@@ -499,7 +499,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
                            'service': 'service_name',
                            'failurePolicy': {'retry': {}}}},)
     ])
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_valid_trigger_union_field(self, trigger, mock_hook):
         mock_hook.return_value.upload_function_zip.return_value = 'https://uploadUrl'
         mock_hook.return_value.get_function.side_effect = mock.Mock(
@@ -528,7 +528,7 @@ class GcfFunctionDeployTest(unittest.TestCase):
         )
         mock_hook.reset_mock()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_extra_parameter(self, mock_hook):
         mock_hook.return_value.create_new_function.return_value = True
         body = deepcopy(VALID_BODY)
@@ -561,7 +561,7 @@ class GcfFunctionDeleteTest(unittest.TestCase):
         'updateTime': '2018-08-23T00:00:00Z',
         'versionId': '1', 'runtime': 'nodejs6'}
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_delete_execute(self, mock_hook):
         mock_hook.return_value.delete_function.return_value = \
             self._DELETE_FUNCTION_EXPECTED
@@ -577,7 +577,7 @@ class GcfFunctionDeleteTest(unittest.TestCase):
         )
         self.assertEqual(result['name'], self._FUNCTION_NAME)
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_correct_name(self, mock_hook):
         op = GcfFunctionDeleteOperator(
             name="projects/project_name/locations/project_location/functions"
@@ -588,7 +588,7 @@ class GcfFunctionDeleteTest(unittest.TestCase):
         mock_hook.assert_called_once_with(api_version='v1',
                                           gcp_conn_id='google_cloud_default')
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_invalid_name(self, mock_hook):
         with self.assertRaises(AttributeError) as cm:
             op = GcfFunctionDeleteOperator(
@@ -601,7 +601,7 @@ class GcfFunctionDeleteTest(unittest.TestCase):
             FUNCTION_NAME_PATTERN))
         mock_hook.assert_not_called()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_empty_name(self, mock_hook):
         mock_hook.return_value.delete_function.return_value = \
             self._DELETE_FUNCTION_EXPECTED
@@ -614,7 +614,7 @@ class GcfFunctionDeleteTest(unittest.TestCase):
         self.assertEqual(str(err), 'Empty parameter: name')
         mock_hook.assert_not_called()
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_gcf_error_silenced_when_function_doesnt_exist(self, mock_hook):
         op = GcfFunctionDeleteOperator(
             name=self._FUNCTION_NAME,
@@ -629,7 +629,7 @@ class GcfFunctionDeleteTest(unittest.TestCase):
             'projects/project_name/locations/project_location/functions/function_name'
         )
 
-    @mock.patch('airflow.contrib.operators.gcp_function_operator.GcfHook')
+    @mock.patch('airflow.gcp.operators.functions.GcfHook')
     def test_non_404_gcf_error_bubbled_up(self, mock_hook):
         op = GcfFunctionDeleteOperator(
             name=self._FUNCTION_NAME,
